Add alternative extension method to create an HTTP response
